Use @terry/shared for communication
