+++ frontend/packages/communication/src/index.tsx
@@ -5,11 +5,11 @@ import ReactDOM from "react-dom";
 import {
   BrowserRouter as Router, Redirect, Route, Switch,
 } from "react-router-dom";
+import { CommunicationContextProvider } from "@terry/shared/_/hooks/useCommunication";
 import { Announcements } from "./components/Announcements";
 import { Home } from "./components/Home";
 import { Login } from "./components/Login";
 import { Navbar } from "./components/Navbar";
-import { CommunicationContextProvider } from "./hooks/useCommunication";
 import { useLogin } from "./hooks/useLogin";
 
 function App() {
@@ -18,7 +18,7 @@ function App() {
 
   return (
     <>
-      <CommunicationContextProvider>
+      <CommunicationContextProvider token={token}>
         <Navbar />
         <Switch>
           <Route path="/announcements" component={Announcements} />
