--- frontend/packages/communication/src/terry-frontend/Loading.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-import "./Loading.css";
-
-export function Loading() {
-  return (
-    <div className="sk-folding-cube">
-      <div className="sk-cube1 sk-cube" />
-      <div className="sk-cube2 sk-cube" />
-      <div className="sk-cube4 sk-cube" />
-      <div className="sk-cube3 sk-cube" />
-    </div>
-  );
-}
