--- frontend/packages/communication/src/terry-frontend/useTriggerUpdate.hook.ts
@@ -1,17 +1,0 @@
-import { useState } from "react";
-
-/**
- * ```
- * const [dep, trigger] = useTriggerUpdate();
- *
- * useEffect(() => {
- *   // ...
- * }, [dep]);
- *
- * // call trigger() to trigger the useEffect update
- * ```
- */
-export function useTriggerUpdate() {
-  const [count, setCount] = useState(0);
-  return [count, () => setCount(count + 1)] as const;
-}
