+++ frontend/packages/shared/src/components/Announcement.tsx
@@ -1,7 +1,7 @@
 import { DateTime } from "luxon";
 import React from "react";
 import ReactMarkdown from "react-markdown";
-import { AbsoluteDateComponent } from "src/terry-frontend/datetime.views";
+import { AbsoluteDate } from "./AbsoluteDate";
 
 type Props = {
     title: string;
@@ -15,7 +15,7 @@ export function Announcement({
 }: Props) {
   return (
     <div className={`alert alert-${severity}`}>
-      <span className="float-right"><AbsoluteDateComponent clock={() => DateTime.fromJSDate(new Date())} date={date} /></span>
+      <span className="float-right"><AbsoluteDate clock={() => DateTime.fromJSDate(new Date())} date={date} /></span>
       <h5 className="alert-heading">
         {title}
       </h5>
